// src/app/Models/GetOtCauseByForm.ts
// Generated by https://quicktype.io

export interface GetOtCauseByFormDataClass {
    CauseID:       string;
    CauseCode:     string;
    CauseNameC:    string;
    Seq:           number;
    OnCall:        boolean;
    NotDisplay:    boolean;
    NotCalculate:  boolean;
    NotFood:       boolean;
    NotCheckCard:  boolean;
    Holiday:       boolean;
    HolidayTypeID: string;
    RateMasterID:  number;
}